S09P31A608-543 Feat: 랭킹 관련 기능
- 내 랭킹 조회
- 전체 랭킹 조회
- Redis SortedSet에 데이터 삽입
- Redis key에 해당하는 데이터 초기화

+++ backend/src/main/java/com/a608/musiq/domain/member/domain/MemberInfo.java
@@ -38,7 +38,7 @@ public class MemberInfo extends BaseTimeEntity {
 
 	@NotNull
 	@Column
-	private Integer exp;
+	private Double exp;
 
 	@NotNull
 	@Builder.Default
